pyproject.toml
[project]
name = "tractionbuild"
version = "0.2.0"
dependencies = [
  "fastapi>=0.115.0,<0.116.0",
  "streamlit>=1.39.0,<1.40.0",
  "sqlalchemy>=2.0.35,<2.1.0",
  "pydantic>=2.9.2,<2.10.0",
  "opentelemetry-api>=1.27.0,<1.28.0",
  "opentelemetry-sdk>=1.27.0,<1.28.0",
# "opentelemetry-exporter-prometheus>=1.27.0,<1.28.0",#
  "prometheus-client>=0.21.0,<0.22.0",
  "cyclonedx-bom>=4.5.0,<4.6.0",
  "bandit>=1.7.9,<1.8.0",
  "pre-commit>=3.8.0,<3.9.0",
  "ruff>=0.6.8,<0.7.0",
  "black>=24.8.0,<24.9.0",
  "isort>=5.13.2,<5.14.0",
  "mypy>=1.11.2,<1.12.0",
  "pytest>=8.3.3,<8.4.0",
  "pytest-cov>=5.0.0,<5.1.0",
  "sentence-transformers>=3.1.1,<3.2.0", # For semantic caching
  "typer>=0.19.2",
  "crewai>=0.203.1",
  "redis>=6.4.0",
  "python-dotenv>=1.1.1"
]


[tool.ruff]
select = ["E", "F", "W", "I", "B", "S", "TCH"]
line-length = 100

[tool.mypy]
strict = true

[tool.pytest.ini_options]
addopts = "-q -x --cov=src --cov-report=term-missing:skip-covered"

[tool.coverage.run]
source = ["src"]